add -expert mode- DISABLE CSS CACHE option - this just invalidates the cache on every reload -> TODO: change this in such a way that css files are loaded from theme folder instead of the cache folder. (this should improve loading times again)
cache loading / generating is quite complex in nature, it might not be possible to create a proxy switch and load css settings from cached files and hardcoded css from theme files separately without changing a lot of parts in the scripts. some parts of the system are not self explaning; $module_css_cache_active is nowhere set to true (default is false), but true for theme modules?